requirements.txt
agentforge
discord.py